fix cr suggestions
