+++ App.tsx
@@ -1,36 +1,10 @@
-import {
-  ApolloClient,
-  ApolloProvider,
-  concat,
-  HttpLink,
-  InMemoryCache,
-} from '@apollo/client';
-import {setContext} from '@apollo/client/link/context';
-import AsyncStorage from '@react-native-async-storage/async-storage';
+import {ApolloProvider} from '@apollo/client';
 import React from 'react';
 import {Navigation} from 'react-native-navigation';
 import {HomePage} from './src/screens/home/home-page';
 import {LoginPage} from './src/screens/login/login-page';
 import {UserPage} from './src/screens/users/user-page';
-
-const httpLink = new HttpLink({
-  uri: 'https://tq-template-server-sample.herokuapp.com/graphql',
-});
-
-const auth = setContext(async (_, {headers}) => {
-  const token = await AsyncStorage.getItem('token');
-  return {
-    headers: {
-      ...headers,
-      authorization: token ? token : '',
-    },
-  };
-});
-
-export const apolloClient = new ApolloClient({
-  link: concat(auth, httpLink),
-  cache: new InMemoryCache(),
-});
+import {apolloClient} from './src/utils/client';
 
 Navigation.registerComponent('Login', () => props => (
   <ApolloProvider client={apolloClient}>
