+++ src/screens/users/user-page.tsx
@@ -1,7 +1,7 @@
 import {useQuery} from '@apollo/client';
 import React from 'react';
 import {FlatList, Text, View} from 'react-native';
-import {GET_USER} from '../../requests';
+import {GET_USER} from '../../utils/requests';
 import {styleUser} from './user-page.styles';
 
 interface User {
@@ -10,14 +10,16 @@ interface User {
   email: string;
 }
 
-const renderUser = (item: {item: User}) => {
+const renderUser = ({item}: {item: User}) => {
   return (
     <View style={styleUser.container}>
-      <Text style={styleUser.text}>Name: {item.item.name}</Text>
-      <Text style={styleUser.text}>Email: {item.item.email}</Text>
+      <Text style={styleUser.text}>Name: {item.name}</Text>
+      <Text style={styleUser.text}>Email: {item.email}</Text>
     </View>
   );
 };
+
+const keyExtractor = (item: {id: string}) => item.id;
 
 export const UserPage = () => {
   const {data} = useQuery(GET_USER);
@@ -26,7 +28,7 @@ export const UserPage = () => {
       <FlatList
         data={data?.users.nodes}
         renderItem={renderUser}
-        keyExtractor={item => item.id}
+        keyExtractor={keyExtractor}
       />
     </View>
   );
