+++ src/utils/client.ts
@@ -1,0 +1,22 @@
+import {ApolloClient, concat, HttpLink, InMemoryCache} from '@apollo/client';
+import {setContext} from '@apollo/client/link/context';
+import AsyncStorage from '@react-native-async-storage/async-storage';
+
+const httpLink = new HttpLink({
+  uri: 'https://tq-template-server-sample.herokuapp.com/graphql',
+});
+
+const auth = setContext(async (_, {headers}) => {
+  const token = await AsyncStorage.getItem('token');
+  return {
+    headers: {
+      ...headers,
+      authorization: token ? token : '',
+    },
+  };
+});
+
+export const apolloClient = new ApolloClient({
+  link: concat(auth, httpLink),
+  cache: new InMemoryCache(),
+});
